LPS-77183 organizations-service - moves classes to the organizations-service module
